+++ packages/utils/src/utils/asserts.utils.spec.ts
@@ -1,4 +1,5 @@
 import {
+  asNonNullish,
   assertNonNullish,
   assertPercentageNumber,
   InvalidPercentageError,
@@ -33,6 +34,49 @@ describe("asserts-utils", () => {
     it("should not throw an exception if valid object", () => {
       const call = () => assertNonNullish({});
       expect(call).not.toThrow();
+    });
+
+    it("should make value of non-nullable type", () => {
+      const getStringOrNull = (): string | null => "test";
+      const value: string | null = getStringOrNull();
+      assertNonNullish(value);
+      const nonNullValue: string = value;
+    });
+  });
+
+  describe("asNonNullish", () => {
+    it("should throw an exception if undefined", () => {
+      const call = () => asNonNullish(undefined);
+
+      expect(call).toThrow();
+    });
+
+    it("should throw an exception if null", () => {
+      const call = () => asNonNullish(null);
+
+      expect(call).toThrow();
+    });
+
+    it("should throw an exception with particular message", () => {
+      const call = () => asNonNullish(undefined, "Test error");
+
+      expect(call).toThrow(new NullishError("Test error"));
+    });
+
+    it("should not throw an exception if valid primitive type", () => {
+      const call = () => asNonNullish(1);
+      expect(call).not.toThrow();
+    });
+
+    it("should not throw an exception if valid object", () => {
+      const call = () => asNonNullish({});
+      expect(call).not.toThrow();
+    });
+
+    it("should return the value if valid", () => {
+      const value: string | undefined = "test";
+      const result: string = asNonNullish(value);
+      expect(result).toBe(value);
     });
   });