+++ CHANGELOG.md
@@ -23,6 +23,7 @@ The current status of the libraries at the time of the release is as follows:
 - Add optional `includeEmptyNeurons` parameter to `listNeurons`.
 - Extend `eip1559TransactionPrice` for Erc20.
 - Add "Protocol Canister Management" and "Service Nervous System Management" topics support.
+- Add `asNonNullish` function, like `assertNonNullish` but returns the value.
 
 ## Fix